rename canonical metrics to client metrics
